removed study_with_related_records_serializer added an attribute 'with_related_records' which can be set to true
study_serializer adds the related records to the serialized json if with_related_records
is true
